Herramientas

Description

Modulo : Herramientas

Descripcion : Funciones adicionales para manipular el laberinto y las funciones del sabio

Autores : Francisco Marquez 12-11163usb.ve Daniel Francis 12-10863usb.ve

Funciones de construcción

poblarLaberinto

poblarLaberinto :: Laberinto -> String -> Laberinto

Toma un laberinto y le construye una ruta explorable.

paredAbierta

paredAbierta :: Laberinto -> String -> Laberinto

Abre una pared (dirección a Nothing) en el laberinto a partir del resto de un camino dado

derrumbeLaberinto

derrumbeLaberinto :: Laberinto -> [Char] -> Char -> Laberinto

Se deshace del laberinto alcanzable en el destino de una ruta dada

derrumbaPared

derrumbaPared :: Laberinto -> Char -> Laberinto

Derrumba una pared de la trifurcación del laberinto actual

Funciones de validación

rutaValida

rutaValida :: String -> Bool

Verifica que un string de ruta solo contenga i,r ó d

letraValida

letraValida :: Char -> Bool

Verifica que un caracter sea i,r ó d

tesoroLocal

tesoroLocal :: Laberinto -> Maybe Tesoro

Retorna el tesoro del laberinto actual

trifurcacionLocal

trifurcacionLocal :: Laberinto -> Trifurcacion

Retorna la trifurcación de laberinto actual

Otras funciones

clear

clear :: IO ()

Limpia la pantalla