Laberinto

Description

Modulo : Laberinto

Descripcion : Constructores y funciones de acceso para el tipo de dato abstracto Laberinto.

Autores : Francisco Marquez 12-11163usb.ve Daniel Francis 12-10863usb.ve

Tipos de dato o clases

Laberinto

data Laberinto

Tipo de dato laberinto, con una trifurcacion

Constructors

Lab Trifurcacion (Maybe Tesoro)

Trifurcacion

data Trifurcacion

Una trifurcacion que dirije a hasta tres laberintos

Constructors

Tri (Maybe Laberinto) (Maybe Laberinto) (Maybe Laberinto)

Tesoro

data Tesoro

Un tesoro. Puede ser pasado por alto para seguir otro laberinto adyacente

Constructors

Tes String (Maybe Laberinto)

Funciones de Construccion

sinSalida

sinSalida :: Trifurcacion

Trifurcacion sin salida

laberintoNuevo

laberintoNuevo :: Laberinto

Laberinto sin salida

pasarPorAlto

pasarPorAlto :: String -> Maybe Laberinto -> Tesoro

Dado un laberinto y una descripcion, retorna un tesoro

pasarPorAlto'

pasarPorAlto' :: Tesoro -> Laberinto

Retorna el laberinto que sigue despues del tesoro

relacionaCamino

relacionaCamino :: Trifurcacion -> Laberinto -> Char -> Trifurcacion

Relaciona tomar una direccion en una trifurcacion que llega a un laberinto

Funciones de acceso

seguirRuta

seguirRuta :: Laberinto -> String -> Laberinto

Sigue recursivamente una ruta de la forma "ridir" dentro de un laberinto para llegar a otro. Tome "i " como izquierda, "r" como recto y "d" como derecha

izquierda

izquierda :: Laberinto -> Laberinto

Retorna el laberinto que queda al entrar por la izquierda a la trifurcacion

recto

recto :: Laberinto -> Laberinto

Retorna el laberinto que queda al entrar por el medio a la trifurcacion

derecha

derecha :: Laberinto -> Laberinto

Retorna el laberinto que queda al entrar por la derecha a la trifurcacion