Sabio

Description

Modulo : Sabio

Descripcion : Cliente para manejo del laberinto

Autores : Francisco Marquez 12-11163usb.ve Daniel Francis 12-10863usb.ve

Funciones principales del cliente

opciones

opciones :: Laberinto -> IO b

Muestra las opciones disponibles

seleccion

seleccion :: [Char] -> Laberinto -> IO b

Ejecuta la opcion resultante de opciones

Opciones del menú principal

opcion0

opcion0 :: IO a

Salir del programa

opcion1

opcion1 :: IO b

Comenzar a hablar de un laberinto nuevo

opcion2

opcion2 :: Laberinto -> IO b

Preguntar ruta

opcion3

opcion3 :: Laberinto -> IO b

Reportar pared abierta. Se recibe un camino, se recorre el camino hasta alcanzar una pared (un Nothing). La ruta dada a partir de ese momento se convierte en el laberinto alcanzable por esa dirección.

opcion4

opcion4 :: Laberinto -> IO b

Reportar derrumbe. Se sigue la ruta en el laberinto y se elimina el laberinto correspondientea la dirección dada en ese sitio.

opcion5

opcion5 :: Laberinto -> IO b

Reportar tesoro tomado al final de la ruta dada

opcion6

opcion7

opcion7 :: Laberinto -> IO b

Dar nombre al laberinto (Guarda el laberinto como archivo de texto simple)

opcion8

opcion8 :: IO b

Hablar de un laberinto de nombre conocido (Carga un laberinto a partir de un archivo de texto simple, siguiendo el formato impuesto por la opcion7)